PoC integration with FSCHateoasBundle
this is just a proof of concept to show that its possible to leverage the route collection data to add the hateoas link relations.

a proper implementation would add the metadata during the bootstrap phase of Symfony2. this would prevent the runtime overhead and would ensure that the metadata isnt just linked to specific routes/controllers.